requirements.txt
streamlit
openai
python-dotenv
SpeechRecognition
PyAudio
pyttsx3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib